Add BaseRestAssured request spec Add get & delete requests

+++ src/main/kotlin/models/Pet.kt
@@ -3,7 +3,7 @@ package models
 data class Pet(
 	val photoUrls: List<String?>? = null,
 	val name: String? = null,
-	val id: Int? = null,
+	val id: Long? = null,
 	val category: Category? = null,
 	val tags: List<TagsItem?>? = null,
 	val status: String? = null
@@ -11,11 +11,11 @@ data class Pet(
 
 data class Category(
 	val name: String? = null,
-	val id: Int? = null
+	val id: Long? = null
 )
 
 data class TagsItem(
 	val name: String? = null,
-	val id: Int? = null
+	val id: Long? = null
 )
